Update timing API. Update context singleton getters to return references. Fix cube map internal handle being 32 bit instead of 64. Remove poelo and datamanager related files.

+++ Engine/LeopphEngine/src/rendering/Skybox.cpp
@@ -40,7 +40,7 @@ namespace leopph
 
 
 
-	u32 Skybox::get_internal_handle() const
+	u64 Skybox::get_internal_handle() const
 	{
 		return mCubemapHandle;
 	}
